Refactor transcript ingestion handler by provider
